+++ examples/src/bin/tlsserver-mio.rs
@@ -633,7 +633,7 @@ fn make_config(args: &Args) -> Arc<rustls::ServerConfig<Ring>> {
     }
 
     if args.flag_tickets {
-        config.ticketer = rustls::Ticketer::new::<Ring>().unwrap();
+        config.ticketer = rustls::Ticketer::new().unwrap();
     }
 
     config.alpn_protocols = args
