+++ rustls/examples/internal/bench.rs
@@ -322,7 +322,7 @@ fn make_server_config(
     if resume == ResumptionParam::SessionID {
         cfg.session_storage = ServerSessionMemoryCache::new(128);
     } else if resume == ResumptionParam::Tickets {
-        cfg.ticketer = Ticketer::new::<Ring>().unwrap();
+        cfg.ticketer = Ticketer::new().unwrap();
     } else {
         cfg.session_storage = Arc::new(NoServerSessionStorage {});
     }
