+++ rustls/src/crypto/mod.rs
@@ -1,5 +1,4 @@
 use crate::rand::GetRandomFailed;
-use crate::server::ProducesTickets;
 use crate::{Error, NamedGroup};
 
 use std::fmt::Debug;
@@ -11,9 +10,6 @@ pub mod ring;
 pub trait CryptoProvider: Send + Sync + 'static {
     /// KeyExchange operations that are supported by the provider.
     type KeyExchange: KeyExchange;
-
-    /// Build a ticket generator.
-    fn ticket_generator() -> Result<Box<dyn ProducesTickets>, GetRandomFailed>;
 
     /// Fill the given buffer with random bytes.
     fn fill_random(buf: &mut [u8]) -> Result<(), GetRandomFailed>;
